pyproject.toml
[project]
name = "turtwig"
version = "0.1.0"
description = "Utility functions and starter code for medical data analysis."
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "black>=24.10.0",
    "cytoolz>=1.0.1",
    "fn-py>=0.6.0",
    "h5py>=3.12.1",
    "isort>=5.13.2",
    "loguru>=0.7.3",
    "medpy>=0.5.2",
    "nibabel>=5.3.2",
    "pathos>=0.3.3",
    "pydicom>=3.0.1",
    "pytest>=8.3.4",
    "pytest-mock>=3.14.0",
    "rt-utils>=1.2.7",
    "simpleitk>=2.4.0",
    "tqdm>=4.67.1",
]
